refactor(cmd): extract commandRouter into its own file
cmd/cmd.go had grown to ~430 lines and was doing three jobs at once:
GitDeps interface + Cmd struct + NewCmd wiring, and also the handler
map / router / registry coverage check. Move the router-related code
(commandRouter type, newCommandRouter, route, missingHandlers) to a
new cmd/router.go file.

Pure code motion: no behavior change, no test change. Each line was
moved verbatim apart from adding doc comments on the now-exported
API surface (within the package).

cmd.go:   430 -> 351 lines
router.go:       101 lines (new)

Motivation: future work on the handler map (per-command help routing,
alias expansion, telemetry) is a focused change in a small file
rather than a diff-noise-heavy change in the big one.
